requirements.txt
absl-py==0.9.0
aiofiles==0.4.0
aiohttp==3.6.2
aioredis==1.3.1
alabaster==0.7.10
aliyun-python-sdk-core==2.13.16
aliyun-python-sdk-ecs==4.19.4
aliyun-python-sdk-slb==3.2.17
aliyun-python-sdk-sts==3.0.2
amqp==1.4.9
aniso8601==7.0.0
ansible==2.9.10
anyjson==0.3.3
appdirs==1.4.4
APScheduler==3.6.3
arrow==0.15.5
asgiref==3.2.10
asn1crypto==0.24.0
astor==0.8.1
astroid==2.2.5
astropy==3.0.2
async-timeout==3.0.1
atomicwrites==1.4.0
attrs==19.3.0
autobahn==20.4.3
Automat==20.2.0
autopep8==1.5.3
Babel==2.5.3
backcall==0.1.0
backports.csv==1.0.7
backports.shutil-get-terminal-size==1.0.0
bcrypt==3.1.7
beautifulsoup4==4.6.0
billiard==3.3.0.23
bitarray==0.8.1
bkcharts==0.2
blaze==0.10.1
bleach==3.1.4
blinker==1.4
bokeh==0.12.16
boto==2.48.0
Bottleneck==1.3.0
cachetools==4.0.0
celery==3.1.26.post2
certifi==2020.4.5.2
cffi==1.14.0
channels==2.2.0
channels-redis==2.4.2
chardet==3.0.4
click==6.7
cloudpickle==0.5.3
clyent==1.2.1
colorama==0.4.2
comtypes==1.1.4
constantly==15.1.0
contextlib2==0.5.5
contextvars==2.4
coreapi==2.3.3
coreschema==0.0.4
coverage==5.0
cryptography==2.9.2
cycler==0.10.0
Cython==0.28.2
cytoolz==0.9.0.1
daphne==2.5.0
dask==0.17.5
datashape==0.5.2
decorator==4.3.0
defer==1.0.4
defusedxml==0.6.0
dictdiffer==0.8.1
diff-match-patch==20181111
distlib==0.3.1
distributed==1.21.8
distro==1.4.0
Django==2.2.14
django-admin-list-filter-dropdown==1.0.2
django-apscheduler==0.3.0
django-braces==1.14.0
django-celery==3.3.1
django-ckeditor==5.9.0
django-cors-headers==3.0.2
django-crispy-forms==1.9.1
django-filter==2.2.0
django-filters==0.2.1
django-formtools==2.2
django-haystack==2.8.1
django-import-export==1.2.0
django-inline-actions==2.2.0
django-js-asset==1.2.2
django-mysql==3.5.0
django-oauth-toolkit==1.3.2
django-otp==0.9.1
django-phonenumber-field==3.0.1
django-redis==4.12.1
django-rest-framework-social-oauth2==1.1.0
django-rest-swagger==2.2.0
django-reversion==3.0.7
django-simple-history==2.10.0
django-simpleui==4.0.2
django-two-factor-auth==1.12
djangorestframework==3.10.1
djangorestframework-jwt==1.11.0
dnspython==1.16.0
docutils==0.14
dramatiq==1.9.0
drf-haystack==1.8.6
ecdsa==0.15
elasticsearch==7.7.1
entrypoints==0.2.3
et-xmlfile==1.0.1
fabric2==2.5.0
Faker==4.0.1
fastcache==1.0.2
fasteners==0.14.1
filelock==3.0.4
flake8==3.8.3
Flask==1.0.2
Flask-Cors==3.0.4
funcy==1.14
future==0.18.2
gast==0.2.2
gevent==1.3.0
gitlab==1.0.2
glob2==0.6
google-auth==1.11.0
google-auth-oauthlib==0.4.1
google-pasta==0.1.8
graphene==2.1.8
graphene-django==2.10.1
graphql-core==2.3.2
graphql-relay==2.0.1
greenlet==0.4.13
grpcio==1.26.0
h11==0.8.1
h2==3.1.1
h5py==2.10.0
HeapDict==1.0.0
helpdev==0.7.1
hiredis==1.0.1
hpack==3.0.0
hstspreload==2020.1.30
html5lib==1.0.1
httplib2==0.18.1
httptools==0.0.13
httpx==0.11.1
hyperframe==5.2.0
hyperlink==19.0.0
icc-rt==2020.0.133
idna==2.6
idna-ssl==1.1.0
imageio==2.3.0
imagesize==1.0.0
immutables==0.11
importlib-metadata==1.5.0
importlib-resources==3.0.0
incremental==17.5.0
intel-openmp==2020.0.133
intervaltree==3.0.2
invoke==1.4.1
ipykernel==5.3.4
ipython==6.4.0
ipython-genutils==0.2.0
ipywidgets==7.2.1
isort==4.3.21
itsdangerous==0.24
itypes==1.2.0
jdcal==1.4.1
jedi==0.17.1
jeepney==0.4.3
Jinja2==2.11.2
jmespath==0.10.0
joblib==0.16.0
jsonschema==2.6.0
jupyter-client==6.1.6
jupyter-core==4.6.3
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
keyring==18.0.1
kiwisolver==1.0.1
kombu==3.0.37
kubernetes==11.0.0
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lazy-object-proxy==1.4.1
llvmlite==0.33.0
locket==0.2.0
lockfile==0.12.2
louis==1.3
luigi==3.0.1
lxml==4.5.2
macaroonbakery==1.3.1
Mako==1.1.0
Markdown==3.1.1
MarkupPy==1.14
MarkupSafe==1.0
matplotlib==3.3.0
mccabe==0.6.1
mistune==0.8.3
mkl==2019.0
mongoengine==0.19.1
monotonic==1.5
more-itertools==4.1.0
mpmath==1.0.0
msgpack==0.6.2
msgpack-python==0.5.6
multi-key-dict==2.0.3
multidict==4.7.4
multipledispatch==0.5.0
mysql-connector-python==8.0.19
mysqlclient==1.4.6
nbconvert==5.3.1
nbformat==4.4.0
netifaces==0.10.4
networkx==2.1
nltk==3.4.5
nmap==0.0.1
nose==1.3.7
notebook==5.7.8
numba==0.50.1
numexpr==2.6.5
numpy==1.18.1
numpydoc==0.8.0
oauthlib==3.1.0
odfpy==1.4.0
odo==0.5.0
olefile==0.45.1
openapi-codec==1.3.2
opencv-python==4.2.0.32
openpyxl==2.6.2
opt-einsum==3.1.0
packaging==17.1
pandas==0.22.0
pandocfilters==1.4.2
paramiko==2.6.0
paramiko-expect==0.2.8
parso==0.7.0
partd==0.3.8
path.py==11.0.1
pathlib2==2.3.2
pathtools==0.1.2
patsy==0.5.0
pbr==5.4.5
pep8==1.7.1
pexpect==4.8.0
phonenumbers==8.12.4
pickleshare==0.7.4
Pillow==7.0.0
pinyin==0.4.0
pipenv==2018.11.26
pkginfo==1.4.2
pluggy==0.6.0
plumbum==1.6.8
ply==3.11
prometheus-client==0.8.0
promise==2.3
prompt-toolkit==1.0.15
protobuf==3.11.2
psutil==5.6.6
ptyprocess==0.6.0
py==1.5.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycosat==0.6.3
pycparser==2.18
pycrypto==2.6.1
pycryptodome==3.9.8
pydocstyle==5.0.2
pyflakes==2.2.0
Pygments==2.2.0
PyHamcrest==2.0.2
PyJWT==1.7.1
pylint==2.3.1
pymacaroons==0.13.0
pymongo==3.10.1
PyMySQL==0.9.3
PyNaCl==1.4.0
PyOpenGL==3.1.5
pyOpenSSL==18.0.0
pyparsing==2.4.7
pypinyin==0.38.0
PyQt5==5.12.3
PyQt5-sip==12.8.0
PyQtWebEngine==5.12.1
pyRFC3339==1.1
PySocks==1.6.8
pysolr==3.9.0
pytest==3.5.1
pytest-arraydiff==0.2
pytest-astropy==0.3.0
pytest-doctestplus==0.1.3
pytest-openfiles==0.3.0
pytest-remotedata==0.2.1
python-daemon==2.1.2
python-dateutil==2.8.0
python-debian==0.1.36
python-gitlab==2.2.0
python-jenkins==1.7.0
python-jsonrpc-server==0.3.4
python-language-server==0.34.1
python3-openid==3.1.0
pytz==2019.3
PyWavelets==1.1.1
pyxdg==0.26
PyYAML==5.3.1
pyzmq==19.0.1
QDarkStyle==2.8.1
qrcode==6.1
QtAwesome==0.7.2
qtconsole==4.7.5
QtPy==1.9.0
raven==6.10.0
redis==3.5.3
reportlab==3.5.34
requests==2.22.0
requests-oauthlib==1.3.0
requests-unixsocket==0.2.0
reversion==0.2
rfc3986==1.3.2
rope==0.11.0
rsa==4.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
Rx==1.6.1
sanic==20.6.3
scikit-image==0.17.2
scikit-learn==0.23.1
scipy==1.5.1
seaborn==0.10.1
SecretStorage==2.3.1
Send2Trash==1.5.0
service-identity==18.1.0
simplegeneric==0.8.1
simplejson==3.16.0
singledispatch==3.4.0.3
sip==5.3.0
six==1.14.0
sniffio==1.1.0
snowballstemmer==2.0.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
sortedcollections==1.2.1
sortedcontainers==2.2.2
Sphinx==3.1.2
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
sphinxcontrib-websupport==1.2.3
spyder==4.1.4
spyder-kernels==1.9.2
SQLAlchemy==1.3.18
sqlparse==0.3.1
statsmodels==0.11.1
sympy==1.1.1
tables==3.6.1
tablib==0.13.0
tbb==2019.0
tblib==1.3.2
tencentcloud-sdk-python==3.0.200
tensorboard==2.0.2
tensorflow-estimator==2.0.1
termcolor==1.1.0
terminado==0.8.1
testpath==0.3.1
testresources==2.0.1
text-unidecode==1.2
threadpoolctl==2.1.0
tifffile==2020.7.22
timeout-decorator==0.4.1
toml==0.10.1
toolz==0.9.0
tornado==5.1.1
tqdm==4.46.1
traitlets==4.3.2
twilio==6.42.0
Twisted==20.3.0
txaio==20.4.1
typed-ast==1.4.0
typing==3.6.4
typing-extensions==3.7.4.1
tzlocal==2.1
ujson==3.0.0
unicodecsv==0.14.1
uritemplate==3.0.1
urllib3==1.25.9
uvloop==0.14.0
vine==1.3.0
virtualenv==16.7.9
virtualenv-clone==0.5.3
wadllib==1.3.3
watchdog==0.10.3
wcwidth==0.1.7
webencodings==0.5.1
websocket-client==0.57.0
websockets==5.0.1
Werkzeug==0.15.3
Whoosh==2.7.4
widgetsnbextension==3.2.1
win-inet-pton==1.0.1
win-unicode-console==0.5
wincertstore==0.2
wrapt==1.12.1
wurlitzer==2.0.1
xlrd==1.2.0
XlsxWriter==1.2.9
xlwt==1.3.0
xmltodict==0.12.0
yapf==0.30.0
yarl==1.4.2
zict==2.0.0
zipp==3.1.0
zope.interface==5.1.0
